University Schedule Gateway Python/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
psycopg[binary,pool]==3.1.19
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
pydantic-settings==2.4.0
